requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asyncio==3.4.3
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
combo_lock==0.2.6
discord.py==2.4.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.10.0
huggingface-hub==0.26.0
idna==3.10
Jinja2==3.1.4
json_database==0.7.0
m3u8==6.0.0
MarkupSafe==3.0.2
memory-tempfile==2.2.3
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.1
numpy==2.1.2
packaging==24.1
pafy==0.5.5
propcache==0.2.0
pycountry==24.6.1
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
setuptools==75.2.0
soupsieve==2.6
sympy==1.13.1
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
urllib3==2.2.3
wheel==0.44.0
yarl==1.15.5
youtube-dl==2021.12.17
